% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{iv}
\alias{iv}
\title{Independend Validation (IV) using 'rsample'}
\usage{
iv(data, m = 20, ...)
}
\arguments{
\item{data}{A data frame.}

\item{m}{The number of samples in the first assesment set. The current
standard value is 20, adapt to your needs.}

\item{...}{Not currently used.}
}
\value{
A tibble with classes \code{iv}, \code{rset}, \code{tbl_df},
\code{tbl}, and \code{data.frame}. The results include a column for the data
split objects and one identification variable. Tries to model the approach of
e.g. \code{\link[rsample]{vfold_cv}}.
}
\description{
Independend Validation (IV) takes an initial subset of size \code{m}
of the data for an analysis. For the assessment a single case of the data is
taken. This case is consequently included in the analysis set. This procedure
is repeated till all but one cases are included in the analysis set.
}
\examples{
iv(mtcars)
iv(mtcars, m = 10)

}
